refactor custom methods implementation
