+++ src/config.py
@@ -1,17 +1,27 @@
 import json
-import typing
+from importlib import import_module
 from os import getenv
 from pathlib import Path
 
 from dotenv import load_dotenv
 
-from src.pql.custom_methods import MyAdd
-from src.pql.sklearn import Sklearn
+from src.pql.custom_methods.my_add import MyAdd
+from src.pql.custom_methods.sklearn import Sklearn
 
 __version__ = "0.1.0"
 
 # Load the main configuration file
 load_dotenv(".env")
+
+
+def parse_and_import_custom_methods(import_specs: list) -> dict:
+    return {
+        custom_method.PQL_IDENTIFIER: custom_method
+        for custom_method in [
+            getattr(import_module(import_spec[0]), import_spec[1])
+            for import_spec in [import_spec.split(":") for import_spec in import_specs]
+        ]
+    }
 
 
 class Config:
@@ -50,7 +60,11 @@ class Config:
     ORACLE_CONTRACT_ABI = json.load(open("src/data/oracle_abi.json"))
 
     # User defined custom methods
-    PQL_CUSTOM_METHODS = {MyAdd.PQL_IDENTIFIER: MyAdd, Sklearn.PQL_IDENTIFIER: Sklearn}
+    PQL_CUSTOM_METHODS_SPECS = [
+        "src.pql.custom_methods.my_add:MyAdd",
+        "src.pql.custom_methods.sklearn:Sklearn",
+    ]
+    PQL_CUSTOM_METHODS = parse_and_import_custom_methods(PQL_CUSTOM_METHODS_SPECS)
 
     def __init__(self):
         Path(self.DATA_FOLDER).mkdir(exist_ok=True)
