+++ src/pql/custom_methods/my_add.py
@@ -1,19 +1,6 @@
 import typing
-from abc import abstractmethod
 
-from src.pql.exceptions import PqlCustomMethodNotImplementedError
-
-
-class CustomMethod:
-    PQL_IDENTIFIER: str
-    SCHEMA: dict
-
-    @classmethod
-    @abstractmethod
-    def execute(cls, step: dict, index: int, pipeline) -> typing.Any:
-        raise PqlCustomMethodNotImplementedError(
-            f"Custom method not implemented for {cls.PQL_IDENTIFIER}"
-        )
+from src.pql.custom_methods import CustomMethod
 
 
 class MyAdd(CustomMethod):
